ML_old/requirement.txt
absl-py==1.4.0
alabaster==0.7.10
altair==4.1.0
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
asgiref==3.4.1
asn1crypto==0.24.0
astor==0.8.1
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
blinker==1.5
blis==0.7.11
bokeh==0.12.16
boto==2.48.0
bottle==0.12.25
bottle-websocket==0.2.9
Bottleneck==1.2.1
cachetools==4.2.4
catalogue==2.0.10
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==8.0.4
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
commonmark==0.9.1
comtypes==1.1.4
conda==4.5.4
conda-build==3.10.5
conda-verify==2.0.0
confection==0.1.5
conllu==4.5.3
contextlib2==0.5.5
contextvars==2.4
cryptography==2.2.2
cycler==0.10.0
cymem==2.0.11
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
dataclasses==0.8
datashape==0.5.4
decorator==4.3.0
distlib==0.3.8
distributed==1.21.8
Django==3.2.7
dlib==19.24.2
docutils==0.14
Eel==0.14.0
entrypoints==0.2.3
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Faker==14.2.1
fastcache==1.0.2
filelock==3.4.1
Flask==1.0.2
Flask-Cors==3.0.4
future==1.0.0
gast==0.6.0
gevent==1.3.0
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.18
glob2==0.6
google-search-results==2.4.2
googlesearch-python==1.0.1
googletrans==4.0.0rc1
greenlet==0.4.13
grpcio==1.48.2
h11==0.9.0
h2==3.2.0
h5py==2.7.1
heapdict==1.0.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.0.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.6
imageio==2.3.0
imagesize==1.0.0
imbalanced-learn==0.8.1
imblearn==0.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==6.4.0
ipython_genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==3.0.3
joblib==1.1.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
langcodes==3.3.0
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
mock==5.1.0
more-itertools==4.1.0
mpmath==1.0.0
msgpack-python==0.5.6
multipledispatch==0.5.0
murmurhash==1.0.12
mysql-connector==2.2.9
mysqlclient==2.0.3
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nmslib==2.1.1
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.19.5
numpydoc==0.8.0
oauthlib==3.2.2
odo==0.5.1
olefile==0.45.1
opencv-python==4.5.5.64
openpyxl==2.5.3
packaging==21.3
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
pathy==0.10.3
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==8.4.0
pkginfo==1.4.2
platformdirs==2.4.0
pluggy==0.6.0
ply==3.11
preshed==3.0.9
prompt-toolkit==1.0.15
protobuf==3.19.6
psutil==5.4.5
py==1.5.3
pyarrow==6.0.1
pybind11==2.6.1
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydantic==1.9.2
pydeck==0.6.2
pyflakes==1.6.0
Pygments==2.14.0
pylint==1.8.4
pymongo==4.1.1
Pympler==1.1
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pysbd==0.3.4
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
python-dotenv==0.20.0
pytz==2018.4
pytz-deprecation-shim==0.1.0.post0
PyWavelets==0.5.2
pywin32==305
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
regex==2023.8.8
requests==2.25.1
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==12.6.0
rope==0.10.7
ruamel_yaml==0.15.35
sacremoses==0.0.53
scikit-image==0.13.1
scikit-learn==0.24.2
scipy==1.1.0
scispacy==0.5.3
seaborn==0.8.1
selenium==3.141.0
semver==2.13.0
Send2Trash==1.5.0
sentence-transformers==1.2.1
sentencepiece==0.1.91
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
sqlparse==0.4.4
srsly==2.4.8
statsmodels==0.9.0
streamlit==1.10.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
textblob==0.17.1
thinc==8.1.12
threadpoolctl==3.1.0
tokenizers==0.9.3
toml==0.10.2
toolz==0.9.0
torch==1.10.2
torchvision==0.11.3
tornado==5.0.2
tqdm==4.64.1
traitlets==4.3.2
transformers==3.5.1
tweepy==3.10.0
typer==0.9.4
typing==3.6.4
typing_extensions==4.1.1
tzdata==2025.1
tzlocal==4.2
unicodecsv==0.14.1
urllib3==1.22
validators==0.20.0
virtualenv==20.17.1
wasabi==1.1.3
watchdog==2.3.1
wcwidth==0.1.7
webdriver-manager==3.7.1
webencodings==0.5.1
Werkzeug==0.14.1
whichcraft==0.6.1
whois==0.9.13
widgetsnbextension==3.2.1
win-unicode-console==0.5
win_inet_pton==1.0.1
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
zict==0.1.3
zipp==3.6.0